requirements.txt
streamlit
openai
pandas
matplotlib
plotly
google-generativeai==0.3.2
